+++ packages/kontinuous/src/cli/commands/config.js
@@ -19,6 +19,7 @@ module.exports = (program) =>
     .addOption(options.private)
     .addOption(options.deployKey)
     .addOption(options.gitDiffEnabled)
+    .addOption(options.gitOrg)
     .option(
       "--remote",
       "select config using kontinuous config from remote repo"
@@ -91,5 +92,5 @@ module.exports = (program) =>
         outputValue = value || ""
       }
 
-      process.stdout.write(outputValue)
+      process.stdout.write(outputValue.toString())
     })