+++ plugins/contrib/patches/log.js
@@ -1,26 +1,41 @@
 module.exports = async (manifests, options, { utils }) => {
-  const { enabled: enabledDefault = true } = options
+  const { enabled: enabledDefault } = options
+
+  const enabledByKind = {
+    Statefulset: false,
+    Deployment: false,
+    Job: true,
+    ...(options.enabledByKind || {}),
+  }
+
   const { kindIsRunnable } = utils
 
   for (const manifest of manifests) {
     const pluginLog = manifest.metadata?.annotations?.["kontinuous/plugin.log"]
 
-    if (!kindIsRunnable(manifest.kind)) {
+    const { kind } = manifest
+    if (!kindIsRunnable(kind)) {
       if (pluginLog !== undefined) {
         delete manifest.metadata.annotations["kontinuous/plugin.log"]
       }
       continue
     }
 
-    if (pluginLog === undefined) {
-      if (!manifest.metadata) {
-        manifest.metadata = {}
-      }
-      if (!manifest.metadata.annotations) {
-        manifest.metadata.annotations = {}
-      }
-      manifest.metadata.annotations["kontinuous/plugin.log"] =
-        enabledDefault.toString()
+    if (pluginLog !== undefined) {
+      continue
     }
+
+    if (!manifest.metadata) {
+      manifest.metadata = {}
+    }
+    if (!manifest.metadata.annotations) {
+      manifest.metadata.annotations = {}
+    }
+
+    const logEnabled =
+      enabledByKind[kind] !== undefined ? enabledByKind[kind] : enabledDefault
+
+    manifest.metadata.annotations["kontinuous/plugin.log"] =
+      logEnabled.toString()
   }
 }