+++ plugins/fabrique/values-compilers/index.js
@@ -1,7 +1,6 @@
 module.exports = [
   // "../charts/contrib/values-compilers",
   "../charts/contrib/values-compilers/01-unfold-charts",
-  "../charts/contrib/values-compilers/02-dash-instances",
   "../charts/contrib/values-compilers/03-resolve-alias-of",
   "../charts/contrib/values-compilers/04-implicit-enabled",
   "../charts/contrib/values-compilers/05-jobs",